feat(gallery): implement slide left animation
